NOJIRA bruk fagsak-backend og slå opp personinfo når trengs
