+++ dokumentlager/domene/src/main/java/no/nav/foreldrepenger/melding/behandling/Behandling.java
@@ -8,7 +8,7 @@ import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-import no.nav.foreldrepenger.melding.fagsak.Fagsak;
+import no.nav.foreldrepenger.melding.fagsak.FagsakBackend;
 import no.nav.foreldrepenger.melding.geografisk.Språkkode;
 import no.nav.foreldrepenger.melding.kodeverk.kodeverdi.BehandlingÅrsakType;
 
@@ -28,32 +28,15 @@ public class Behandling {
     private boolean toTrinnsBehandling;
     private String behandlendeEnhetNavn;
     private String ansvarligBeslutter;
-    private Fagsak fagsak;
+    private FagsakBackend fagsakBackend;
     private BehandlingStatus status;
     private String endretAv;
     private Språkkode språkkode;
     private LocalDate originalVedtaksDato;
 
-    private Behandling(Builder builder) {
-        id = builder.id;
-        uuid = builder.uuid;
-        behandlendeEnhetNavn = builder.behandlendeEnhetNavn;
-        behandlingsresultat = builder.behandlingsresultat;
-        resourceLinker = builder.resourceLinker;
-        behandlingType = builder.behandlingType;
-        behandlingstidFristUker = builder.behandlingstidFristUker;
-        opprettetDato = builder.opprettetDato;
-        avsluttet = builder.avsluttet;
-        behandlingÅrsaker = builder.behandlingÅrsaker;
-        ansvarligSaksbehandler = builder.ansvarligSaksbehandler;
-        toTrinnsBehandling = builder.toTrinnsBehandling;
-        ansvarligBeslutter = builder.ansvarligBeslutter;
-        fagsak = builder.fagsak;
-        status = builder.status;
-        endretAv = builder.endretAv;
-        språkkode = builder.språkkode;
-        originalVedtaksDato = builder.originalVedtaksDato;
+    private Behandling() {
     }
+
 
     public LocalDateTime getAvsluttet() {
         return avsluttet;
@@ -140,8 +123,8 @@ public class Behandling {
         return BehandlingType.KLAGE.equals(getBehandlingType());
     }
 
-    public Fagsak getFagsak() {
-        return fagsak;
+    public FagsakBackend getFagsakBackend() {
+        return fagsakBackend;
     }
 
     public boolean erSaksbehandlingAvsluttet() {
@@ -167,14 +150,14 @@ public class Behandling {
         return endretAv;
     }
 
-    public void leggtilFagsak(Fagsak fagsak) {
-        if (this.fagsak == null) {
-            this.fagsak = fagsak;
+    public void leggtilFagsakBackend(FagsakBackend fagsak) {
+        if (this.fagsakBackend == null) {
+            this.fagsakBackend = fagsak;
         }
     }
 
-    public boolean harFagsak() {
-        return fagsak != null;
+    public boolean harFagsakBackend() {
+        return fagsakBackend != null;
     }
 
     public boolean erManueltOpprettet() {
@@ -188,150 +171,107 @@ public class Behandling {
         return new Behandling.Builder();
     }
 
-    public static Behandling.Builder builder(Behandling copy) {
-        return new Behandling.Builder(copy);
-    }
-
     public static class Builder {
-        private String behandlendeEnhetNavn;
-        private Behandlingsresultat behandlingsresultat;
-        private List<BehandlingResourceLink> resourceLinker = new ArrayList<>();
-
-        private long id;
-        private UUID uuid;
-        private Behandling originalBehandling;
-        private BehandlingType behandlingType;
-        private Integer behandlingstidFristUker;
-        private LocalDateTime opprettetDato;
-        private LocalDateTime avsluttet;
-        private List<BehandlingÅrsak> behandlingÅrsaker = new ArrayList<>();
-        private String ansvarligSaksbehandler;
-        private boolean toTrinnsBehandling;
-        private String ansvarligBeslutter;
-        private Fagsak fagsak;
-        private BehandlingStatus status;
-        private String endretAv;
-        private Språkkode språkkode;
-        private LocalDate originalVedtaksDato;
+        private Behandling kladd;
 
         public Builder() {
-        }
-
-        public Builder(Behandling copy) {
-            this.behandlingsresultat = copy.getBehandlingsresultat();
-            this.resourceLinker = copy.getResourceLinker();
-            this.id = copy.getId();
-            this.uuid = copy.getUuid();
-            this.behandlingType = copy.getBehandlingType();
-            this.behandlingstidFristUker = copy.getBehandlingstidFristUker();
-            this.opprettetDato = copy.getOpprettetDato();
-            this.avsluttet = copy.getAvsluttet();
-            this.behandlingÅrsaker = copy.getBehandlingÅrsaker();
-            this.ansvarligSaksbehandler = copy.getAnsvarligSaksbehandler();
-            this.toTrinnsBehandling = copy.isToTrinnsBehandling();
-            this.behandlendeEnhetNavn = copy.getBehandlendeEnhetNavn();
-            this.ansvarligBeslutter = copy.getAnsvarligBeslutter();
-            this.fagsak = copy.getFagsak();
-            this.status = copy.getStatus();
-            this.endretAv = copy.getEndretAv();
-            this.språkkode = copy.getSpråkkode();
-            this.originalVedtaksDato = copy.getOriginalVedtaksDato();
+            this.kladd = new Behandling();
+            this.kladd.resourceLinker = new ArrayList<>();
+            this.kladd.behandlingÅrsaker = new ArrayList<>();
         }
 
         public Behandling.Builder medBehandlendeEnhetNavn(String behandlendeEnhetNavn) {
-            this.behandlendeEnhetNavn = behandlendeEnhetNavn;
+            this.kladd.behandlendeEnhetNavn = behandlendeEnhetNavn;
             return this;
         }
 
         public Behandling.Builder medBehandlingsresultat(Behandlingsresultat behandlingsresultat) {
-            this.behandlingsresultat = behandlingsresultat;
+            this.kladd.behandlingsresultat = behandlingsresultat;
             return this;
         }
 
         public Behandling.Builder leggTilResourceLink(BehandlingResourceLink resourceLink) {
-            this.resourceLinker.add(resourceLink);
+            this.kladd.resourceLinker.add(resourceLink);
             return this;
         }
 
         public Behandling.Builder medId(long id) {
-            this.id = id;
+            this.kladd.id = id;
             return this;
         }
 
         public Behandling.Builder medUuid(UUID uuid) {
-            this.uuid = uuid;
+            this.kladd.uuid = uuid;
             return this;
         }
 
         public Behandling.Builder medBehandlingType(BehandlingType behandlingType) {
-            this.behandlingType = behandlingType;
+            this.kladd.behandlingType = behandlingType;
             return this;
         }
 
         public Behandling.Builder medBehandlingstidFristUker(Integer behandlingstidFristUker) {
-            this.behandlingstidFristUker = behandlingstidFristUker;
+            this.kladd.behandlingstidFristUker = behandlingstidFristUker;
             return this;
         }
 
         public Behandling.Builder medOpprettetDato(LocalDateTime opprettetDato) {
-            this.opprettetDato = opprettetDato;
+            this.kladd.opprettetDato = opprettetDato;
             return this;
         }
 
         public Behandling.Builder medAvsluttet(LocalDateTime avsluttet) {
-            this.avsluttet = avsluttet;
+            this.kladd.avsluttet = avsluttet;
             return this;
         }
 
         public Behandling.Builder medBehandlingÅrsaker(List<BehandlingÅrsak> behandlingÅrsaker) {
-            this.behandlingÅrsaker.addAll(behandlingÅrsaker);
+            this.kladd.behandlingÅrsaker.addAll(behandlingÅrsaker);
             return this;
         }
 
         public Behandling.Builder medAnsvarligSaksbehandler(String ansvarligSaksbehandler) {
-            this.ansvarligSaksbehandler = ansvarligSaksbehandler;
+            this.kladd.ansvarligSaksbehandler = ansvarligSaksbehandler;
             return this;
         }
 
         public Behandling.Builder medToTrinnsBehandling(boolean toTrinnsBehandling) {
-            this.toTrinnsBehandling = toTrinnsBehandling;
+            this.kladd.toTrinnsBehandling = toTrinnsBehandling;
             return this;
         }
 
         public Behandling.Builder medAnsvarligBeslutter(String ansvarligBeslutter) {
-            this.ansvarligBeslutter = ansvarligBeslutter;
+            this.kladd.ansvarligBeslutter = ansvarligBeslutter;
             return this;
         }
 
         public Behandling.Builder medStatus(BehandlingStatus status) {
-            this.status = status;
+            this.kladd.status = status;
             return this;
         }
 
-        public Behandling.Builder medFagsak(Fagsak fagsak) {
-            this.fagsak = fagsak;
+        public Behandling.Builder medFagsakBackend(FagsakBackend fagsak) {
+            this.kladd.fagsakBackend = fagsak;
             return this;
         }
 
         public Behandling.Builder medEndretAv(String endretAv) {
-            this.endretAv = endretAv;
+            this.kladd.endretAv = endretAv;
             return this;
         }
 
         public Behandling.Builder medSpråkkode(Språkkode språkkode) {
-            this.språkkode = språkkode;
+            this.kladd.språkkode = språkkode;
             return this;
         }
 
         public Behandling.Builder medOriginalVedtaksDato(LocalDate originalVedtaksDato) {
-            this.originalVedtaksDato = originalVedtaksDato;
+            this.kladd.originalVedtaksDato = originalVedtaksDato;
             return this;
         }
 
-
-
         public Behandling build() {
-            return new Behandling(this);
+            return this.kladd;
         }
     }
 }
