+++ domenetjenester/brevbestiller/src/main/java/no/nav/foreldrepenger/melding/datamapper/dto/AksjonspunktDtoMapper.java
@@ -3,14 +3,11 @@ package no.nav.foreldrepenger.melding.datamapper.dto;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.enterprise.context.ApplicationScoped;
-
 import no.nav.foreldrepenger.fpsak.dto.behandling.aksjonspunkt.AksjonspunktDto;
 import no.nav.foreldrepenger.melding.aksjonspunkt.Aksjonspunkt;
 import no.nav.foreldrepenger.melding.aksjonspunkt.AksjonspunktDefinisjon;
 import no.nav.foreldrepenger.melding.aksjonspunkt.AksjonspunktStatus;
 
-@ApplicationScoped
 public class AksjonspunktDtoMapper {
 
     static AksjonspunktStatus finnAksjonspunktStatus(String kode) {
@@ -39,7 +36,7 @@ public class AksjonspunktDtoMapper {
         }
     }
 
-    public List<Aksjonspunkt> mapAksjonspunktFraDto(List<AksjonspunktDto> dtoer) {
+    public static List<Aksjonspunkt> mapAksjonspunktFraDto(List<AksjonspunktDto> dtoer) {
         List<Aksjonspunkt> aksjonspunktList = new ArrayList<>();
         for (AksjonspunktDto aksjonspunktDto : dtoer) {
             aksjonspunktList.add(Aksjonspunkt.ny()
