+++ domenetjenester/brevbestiller/src/test/java/no/nav/foreldrepenger/melding/datamapper/brev/InnvilgelseForeldrepengerMapperTest.java
@@ -3,6 +3,7 @@ package no.nav.foreldrepenger.melding.datamapper.brev;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import java.util.Optional;
@@ -16,6 +17,7 @@ import no.nav.foreldrepenger.melding.behandling.Behandling;
 import no.nav.foreldrepenger.melding.datamapper.DomeneobjektProvider;
 import no.nav.foreldrepenger.melding.datamapper.konfig.BrevParametere;
 import no.nav.foreldrepenger.melding.søknad.Søknad;
+import no.nav.foreldrepenger.tps.TpsTjeneste;
 
 public class InnvilgelseForeldrepengerMapperTest {
 
@@ -41,7 +43,7 @@ public class InnvilgelseForeldrepengerMapperTest {
         doReturn(UUID.fromString("EDCD223A-B0F9-4CB4-A9FE-39EFB52A0C50")).when(fjerdeBehandling).getUuid();
         doReturn(UUID.fromString("EECD223A-B0F9-4CB4-A9FE-39EFB52A0C50")).when(femteBehandling).getUuid();
         when(domeneobjektProvider.hentSøknad(Mockito.any())).thenReturn(Optional.empty());
-        mapper = new InnvilgelseForeldrepengerMapper(domeneobjektProvider, brevParametere);
+        mapper = new InnvilgelseForeldrepengerMapper(domeneobjektProvider, brevParametere, mock(TpsTjeneste.class));
     }
 
     @Test
