+++ integrasjon/dtomapper/src/main/java/no/nav/foreldrepenger/melding/dtomapper/DokumentHendelseDtoMapper.java
@@ -1,7 +1,5 @@
 package no.nav.foreldrepenger.melding.dtomapper;
 
-import javax.enterprise.context.ApplicationScoped;
-
 import no.nav.foreldrepenger.melding.behandling.RevurderingVarslingÅrsak;
 import no.nav.foreldrepenger.melding.fagsak.FagsakYtelseType;
 import no.nav.foreldrepenger.melding.hendelser.DokumentHendelse;
@@ -10,42 +8,37 @@ import no.nav.foreldrepenger.melding.kodeverk.kodeverdi.DokumentMalType;
 import no.nav.vedtak.felles.dokumentbestilling.v1.DokumentbestillingV1;
 import no.nav.vedtak.util.StringUtils;
 
-@ApplicationScoped
 public class DokumentHendelseDtoMapper {
 
-    public DokumentHendelseDtoMapper() {
-        //CDI
-    }
-
-    private RevurderingVarslingÅrsak utledRevurderingVarslingsårsak(String varslingsårsak) {
+    private static RevurderingVarslingÅrsak utledRevurderingVarslingsårsak(String varslingsårsak) {
         if (StringUtils.nullOrEmpty(varslingsårsak)) {
             return RevurderingVarslingÅrsak.UDEFINERT;
         }
         return RevurderingVarslingÅrsak.fraKode(varslingsårsak);
     }
 
-    private FagsakYtelseType utledYtelseType(String ytelseType) {
+    private static FagsakYtelseType utledYtelseType(String ytelseType) {
         if (StringUtils.nullOrEmpty(ytelseType)) {
             return null;
         }
         return FagsakYtelseType.fraKode(ytelseType);
     }
 
-    private DokumentMalType utleddokumentMalType(String dokumentmal) {
+    private static DokumentMalType utleddokumentMalType(String dokumentmal) {
         if (StringUtils.nullOrEmpty(dokumentmal)) {
             return null;
         }
         return DokumentMalType.fraKode(dokumentmal);
     }
 
-    private HistorikkAktør utledHistorikkAktør(String historikkAktør) {
+    private static HistorikkAktør utledHistorikkAktør(String historikkAktør) {
         if (StringUtils.nullOrEmpty(historikkAktør)) {
             return HistorikkAktør.UDEFINERT;
         }
         return HistorikkAktør.fraKode(historikkAktør);
     }
 
-    public DokumentHendelse mapDokumentHendelseFraDtoForKafka(DokumentbestillingV1 dokumentbestilling) {
+    public static DokumentHendelse mapDokumentHendelseFraDtoForKafka(DokumentbestillingV1 dokumentbestilling) {
         return DokumentHendelse.builder()
                 .medBehandlingUuid(dokumentbestilling.getBehandlingUuid())
                 .medBestillingUuid(dokumentbestilling.getDokumentbestillingUuid())
